added CQRS to existing database

+++ VoteService/Models/VoteModels.cs
@@ -2,4 +2,4 @@
 
 public record Vote(long Id, long PollId, long UserId, long ChoiceId, DateTime CreatedAt);
 
-public record VoteData(int PollId, int UserId, int ChoiceId);
+public record VoteData(long PollId, long UserId, long ChoiceId);
